fix(contentlayer.config.js): prepend slash to post URL
Thanks Yousuf Elkhoga for the bug report!

This commit updates the `url` computed field in the `Post` document type
within `contentlayer.config.js`. The `resolve` function now prepends a
slash (`/`) to the flattened path of each post, ensuring that all URLs
are absolute paths. This change ensures consistency and correctness in
how URLs are generated for posts across the application.